feat: setup communication domain package
