+++ internal/domain/communication/thread.go
@@ -1,4 +1,4 @@
-package entity
+package communication
 
 import (
 	"slices"
@@ -9,19 +9,17 @@ import (
 
 type ThreadID string
 
-// Thread is a series of ThreadMessages
+// Thread is an entity representing a sequence of messages in a single conversation context.
 type Thread interface {
 	ID() ThreadID
-
-	// Messages returns slice of ThreadMessages.
-	// Each message in the slice is ordered by their id in ascending order and be uniquified.
 	Messages() []ThreadMessage
-
-	// AddMessage adds a ThreadMessage to the Thread.
 	AddMessage(ThreadMessage)
-
-	// AddMessages adds multiple ThreadMessages to the Thread.
 	AddMessages(...ThreadMessage)
+}
+
+type thread struct {
+	id       ThreadID
+	messages []ThreadMessage
 }
 
 func NewThread(id ThreadID) Thread {
@@ -31,28 +29,19 @@ func NewThread(id ThreadID) Thread {
 	}
 }
 
-type thread struct {
-	id       ThreadID
-	messages []ThreadMessage
-}
-
-// AddMessage implements Thread.
 func (t *thread) AddMessage(msg ThreadMessage) {
 	t.AddMessages(msg)
 }
 
-// AddMessages implements Thread.
 func (t *thread) AddMessages(msgs ...ThreadMessage) {
 	t.messages = append(t.messages, msgs...)
 	t.sortMessages()
 }
 
-// ID implements Thread.
 func (t *thread) ID() ThreadID {
 	return t.id
 }
 
-// Messages implements Thread.
 func (t *thread) Messages() []ThreadMessage {
 	return t.messages
 }
