--- internal/domain/event/publisher.go
@@ -1,8 +1,0 @@
-package event
-
-import "context"
-
-type Publisher interface {
-	Subscribe(Kind, Subscriber)
-	Publish(context.Context, Event) error
-}
