--- internal/domain/event/subscriber.go
@@ -1,5 +1,0 @@
-package event
-
-import "context"
-
-type Subscriber func(context.Context, Event) error
